requirements.txt
pymupdf>=1.24.0
supabase>=2.3.4
python-dotenv>=1.0.1
pillow>=10.0.0
tqdm>=4.66.0
